// part9/patientor-backend/src/types.ts
// Domain types for Patientor backend

export interface Diagnosis {
  code: string
  name: string
  latin?: string
}

export enum Gender {
  Male = 'male',
  Female = 'female',
  Other = 'other'
}

// Entries
export interface BaseEntry {
  id: string
  date: string
  specialist: string
  description: string
  diagnosisCodes?: Array<Diagnosis['code']>
}

export interface HospitalEntry extends BaseEntry {
  type: 'Hospital'
  discharge: {
    date: string
    criteria: string
  }
}

export interface OccupationalHealthcareEntry extends BaseEntry {
  type: 'OccupationalHealthcare'
  employerName: string
  sickLeave?: {
    startDate: string
    endDate: string
  }
}

export enum HealthCheckRating {
  Healthy = 0,
  LowRisk = 1,
  HighRisk = 2,
  CriticalRisk = 3
}

export interface HealthCheckEntry extends BaseEntry {
  type: 'HealthCheck'
  healthCheckRating: HealthCheckRating
}

export type Entry = HospitalEntry | OccupationalHealthcareEntry | HealthCheckEntry
export type NewEntry = Omit<Entry, 'id'>

export interface Patient {
  id: string
  name: string
  dateOfBirth: string
  gender: Gender
  occupation: string
  ssn: string
  // eslint-disable-next-line @typescript-eslint/no-empty-object-type
  entries: Entry[]
}

export type NonSensitivePatient = Omit<Patient, 'ssn'>
export type NewPatient = Omit<Patient, 'id'>

// Update NonSensitivePatient to exclude 'entries' as well when returning list
export type NonSensitivePatientListItem = Omit<Patient, 'ssn' | 'entries'>